refactored discovery page
